safari-600.7-branch/LayoutTests/platform/ios-sim/storage/indexeddb/database-name-undefined-expected.txt

Test IndexedDB undefined as record value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

PASS indexedDB.open(); threw exception TypeError: null is not an object (evaluating 'indexedDB.open').
PASS successfullyParsed is true

TEST COMPLETE
